gcc 5

+++ libsrc/compression.cpp
@@ -459,7 +459,7 @@ void compress_image(ISMRMRD::ISMRMRD_Image const &image, std::vector<uint8_t> &b
         auto maximum = std::accumulate(reinterpret_cast<int16 *>(image.data),
                                        reinterpret_cast<int16 *>(image.data) + ismrmrd_size_of_image_data(&image) / ismrmrd_sizeof_data_type(image.head.data_type),
                                        int16(0),
-                                       [](int16 const &a, int16 const &b) { return a | std::abs(b); });
+                                       [](int16 const &a, int16 const &b) { return a | static_cast<int16>(std::abs(b)); });
 
         if (maximum) {
             // Shift until MSB is used for maximum value
